requirements.txt
beautifulsoup4==4.12.3
email_validator==2.2.0
google-ai-generativelanguage==0.6.10
google-api-core==2.23.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-core==2.4.1
google-generativeai==0.8.3
googleapis-common-protos==1.66.0
langchain==0.3.9
langchain-community==0.3.9
langchain-core==0.3.21
langchain-google-calendar-tools==0.0.1
langchain-google-community==2.0.3
langchain-google-genai==2.0.6
langchain-openai==0.2.11
langchain-text-splitters==0.3.2
langgraph==0.2.54
langgraph-checkpoint==2.0.8
langgraph-sdk==0.1.42
langsmith==0.1.147
openai==1.56.1
python-dotenv==1.0.1
pytz==2024.2
streamlit==1.40.2
